Move utils tests to the correct file
