+++ src/utils.test.ts
@@ -1,6 +1,6 @@
 import { PriorityFee } from '@api3/airnode-utilities';
 import { BigNumber } from 'ethers';
-import { parsePriorityFee } from './utils';
+import { parsePriorityFee, calculateTimeout, prepareGoOptions } from './utils';
 
 describe('parsePriorityFee', () => {
   [
@@ -28,5 +28,31 @@ describe('parsePriorityFee', () => {
       const throwingFunction = () => parsePriorityFee(input as PriorityFee);
       expect(throwingFunction).toThrow();
     });
+  });
+});
+
+describe('calculateTimeout', () => {
+  it('calculates the remaining time for timeout', () => {
+    const startTime = 1650548022000;
+    const totalTimeout = 3000;
+    jest.spyOn(Date, 'now').mockReturnValue(1650548023000);
+
+    expect(calculateTimeout(startTime, totalTimeout)).toEqual(2000);
+  });
+});
+
+describe('prepareGoOptions', () => {
+  it('prepares options for the go function', () => {
+    const startTime = 1650548022000;
+    const totalTimeout = 3000;
+    jest.spyOn(Date, 'now').mockReturnValue(1650548023000);
+
+    const expectedGoOptions = {
+      attemptTimeoutMs: 5_000,
+      totalTimeoutMs: 2000,
+      retries: 100_000,
+      delay: { type: 'random' as const, minDelayMs: 0, maxDelayMs: 2_500 },
+    };
+    expect(prepareGoOptions(startTime, totalTimeout)).toEqual(expectedGoOptions);
   });
 });
